% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr.R
\name{b_}
\alias{b_}
\title{Extract Coefficient Point Estimates}
\usage{
b_(v)
}
\arguments{
\item{v}{variable name, unquoted}
}
\description{
Extract Coefficient Point Estimates
}
\details{
After fitting model using regr, use this to extract point estimates of the
coefficients like in Stata.
}
