% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{Generate}
\usage{
gen(...)
}
\arguments{
\item{...}{multiple new variables to add to dataset}
}
\description{
Create new variables to add to your dataset
}
\details{
Add the list of variables in \code{...} to your dataset. Please include names for all your variables, i.e. \code{gen(pe_motheduc = b_(motheduc))} rather than \code{gen(b_motheduc)}. Nothing prevents you from doing that, but the generated variables will have nonsensical names.
}
