% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htv.R
\docType{data}
\name{htv}
\alias{htv}
\title{Wooldridge dataset on educational outcomes. Used for examples throughout the package.}
\format{
An object of class \code{data.frame} with 1230 rows and 23 columns.
}
\usage{
htv
}
\description{
Wooldridge dataset on educational outcomes. Used for examples throughout the package.
}
\keyword{datasets}
