% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr.R
\name{regr}
\alias{regr}
\title{Run a regression}
\usage{
regr(formula)
}
\arguments{
\item{formula}{a formula using variables from the loaded dataset}
}
\description{
Run a regression and store it globally like in Stata. Aims to support as many
features as Stata does (robust standard errors, clustering, confidence levels...)
}
\details{
TODO: add more description to this. It will need to be one of the heavy lifters.
TODO: create a class for this that extends lm functionality
}
\examples{
use(datasets::mtcars)
regr(mpg ~ wt + disp*cyl)

}
