% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{rstata_lm}
\alias{rstata_lm}
\title{Create linear models like Stata}
\usage{
rstata_lm(..., stderr = "homoscedastic", level = 0.95)
}
\arguments{
\item{...}{arguments to lm (see documentation for lm)}

\item{stderr}{method to compute standard errors (name, string or function call). Implemented methods are
"homoscedastic"
"robust"
cluster(group)
"bootstrap"}

\item{level}{alpha level, either a percentage or a number between 0 and 1}
}
\value{
The lm object from R, with vcov matrix and level attached.
}
\description{
This is an \code{lm} call that allows the caller to also specify a covariance structure for the residuals and an alpha level for tests. It's important to use this rather than the standard \code{lm} call, since the rest of the package assumes that the covariance matrix for the betas and the alpha level are attached.

Note that the created object "subclasses" \code{lm}, so you can do anything with the returned object that you would do with a normal linear model object, such as call \code{fitted} and \code{resid}. The returned object also has its own S3 method for vcov, so functions from the stats package like \code{confint} will "work out of the box" and use the covariance matrix created here (rather than the typical homoscedastic case).
}
\examples{

# it works like a regular lm
lm1 <- rstata_lm(mpg ~ wt + cyl, data = datasets::mtcars)

# but also lets you specify a covariance structure
lm2 <- rstata_lm(mpg ~ wt + cyl, data = datasets::mtcars, stderr = robust)

# clustered standard errors are specified as a function call
lm3 <- rstata_lm(mpg ~ wt + cyl, data = datasets::mtcars, stderr = cluster(cyl))

# bootstrapped errors assume 50 replications but can be changed
lm4 <- rstata_lm(mpg ~ wt + cyl, data = datasets::mtcars, stderr = bootstrap)
lm5 <- rstata_lm(mpg ~ wt + cyl, data = datasets::mtcars, stderr = bootstrap(250))

}
