% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnl.R
\name{testnl}
\alias{testnl}
\title{Wald test of multiple hypotheses}
\usage{
testnl(...)
}
\arguments{
\item{...}{one or more (possibly non-linear) expressions}
}
\description{
Test linear and non-linear hypotheses of the coefficients of a fitted regression model.
}
\details{
Jointly tests that the (non-linear) expressions in ... are all equal to zero.
Use a variable name to refer to the coefficient associated with
that variable, for instance \code{testnl(model, 2*age)} to test that
twice the coefficient associated with age is equal to zero.
Separate multiple expressions by commas, i.e. \code{testnl(model, 2*age, motheduc^2)}.
For "weird" variable names, surround in backticks (see example).
}
\examples{
# fit a model with interations for illustration
use(datasets::mtcars)
regr(wt ~ hp*disp)

# first, test that hp = disp
testnl(hp - disp)

# then do a full f test (`hp:disp` is the interaction term)
testnl(hp, disp, `hp:disp`)

# if you want to test the intercept, it's called (Intercept)
testnl(sqrt(`(Intercept)`))

}
