% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Load a dataset for use with rstata}
\usage{
use(dataset = NULL)
}
\arguments{
\item{dataset}{optional, attach a dataframe already in your R workspace}
}
\description{
Load a dataset for use with rstata
}
\details{
Load either a specified dataset or browse your computer for a .dta file
}
\examples{
# suggested usage: browse to a .dta file and use
\dontrun{
use()
}

# alternative usage: re-use a dataframe already in R
use(datasets::mtcars)
}
